EAIOS-BritishPetroleum/data-pipelines/requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-redis==3.6.0
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
